add lenPrefixedWrapper
