+++ frame_metadata/Codec.ts
@@ -1,4 +1,5 @@
 import * as $ from "../deps/scale.ts";
+import { $lenPrefixed } from "../util/$lenPrefixed.ts";
 import { $era } from "./Era.ts";
 import type * as M from "./mod.ts";
 import { TyVisitor } from "./TyVisitor.ts";
@@ -101,6 +102,9 @@ export function DeriveCodec(tys: M.Ty[]): DeriveCodec {
     },
     era() {
       return $era;
+    },
+    lenPrefixedWrapper(_ty, inner) {
+      return $lenPrefixed(this.visit(inner));
     },
     circular(ty) {
       return $.deferred(() => this.cache[ty.id]!);
