+++ mod.ts
@@ -6,6 +6,7 @@ export * as M from "./frame_metadata/mod.ts";
 export { $era, $null, ChainError, type Era } from "./frame_metadata/mod.ts";
 export { kusama, moonbeam, polkadot, rococo, westend } from "./known/mod.ts"; // TODO: get rid of this!
 export {
+  $lenPrefixed,
   type CreateWatchHandler,
   hex,
   mapCreateWatchHandler,
