+++ util/mod.ts
@@ -1,3 +1,4 @@
+export * from "./$lenPrefixed.ts";
 export * from "./branded.ts";
 export * from "./error.ts";
 export * as hex from "./hex.ts";
